## 1.2.0 - RELEASE - 2018-4-9 ### Added - 实现获取目标机进程信息接口 - 实现获取目标机磁盘信息接口 - 实现获取目标机基本信息接口 ### Changed - 重构部署和快照部署的发送部分代码，支持查看实时文件发送状态、进度、速度等信息。 - 丰富部署日志的保存信息。 - 丰富后台打印信息。 - 修改创建、复制组件接口的隔离级别 ### Fixed - 修复部署没有实体文件的组件时数据格式化错误的问题。 - 修复部署日志详情未保存组建信息的问题。 - 修复部署设计绑定组件时判断的逻辑错误。 - 修复删除设备后设备部署信息依然存在的问题。 - 修复工程删除失败的问题 ### Removed - 除去未使用的Redis访问方法

+++ src/main/java/com/rengu/operationsoanagementsuite/Entity/DiskInfoEntity.java
@@ -3,13 +3,13 @@ package com.rengu.operationsoanagementsuite.Entity;
 public class DiskInfoEntity {
 
     private String name;
-    private int size;
-    private int usedSize;
+    private double size;
+    private double usedSize;
 
     public DiskInfoEntity() {
     }
 
-    public DiskInfoEntity(String name, int size, int usedSize) {
+    public DiskInfoEntity(String name, double size, double usedSize) {
         this.name = name;
         this.size = size;
         this.usedSize = usedSize;
@@ -23,19 +23,19 @@ public class DiskInfoEntity {
         this.name = name;
     }
 
-    public int getSize() {
+    public double getSize() {
         return size;
     }
 
-    public void setSize(int size) {
+    public void setSize(double size) {
         this.size = size;
     }
 
-    public int getUsedSize() {
+    public double getUsedSize() {
         return usedSize;
     }
 
-    public void setUsedSize(int usedSize) {
+    public void setUsedSize(double usedSize) {
         this.usedSize = usedSize;
     }
 }
